requirements.txt
# amqp==5.1.1
# # apturl==0.5.2
asgiref==3.5.2
# async-generator==1.10
# async-timeout==4.0.2
# attrs==22.1.0
# backports.zoneinfo==0.2.1
# bcrypt==3.1.7
# billiard==3.6.4.0
# blinker==1.4
# boto3==1.26.12
# botocore==1.29.12
# Brlapi==0.7.0
celery==5.2.7
# certifi==2022.12.7
# chardet==3.0.4
# charset-normalizer==2.1.1
# # chrome-gnome-shell==0.0.0
# click==8.1.3
# click-didyoumean==0.3.0
# click-plugins==1.1.1
# click-repl==0.2.0
# colorama==0.4.3
# # command-not-found==0.3
# coverage==6.5.0
# cryptography==2.8
# # cupshelpers==1.0
# dbus-python==1.2.16
# # defer==1.0.6
# Deprecated==1.2.13
# distro==1.4.0
# # distro-info===0.23ubuntu1
Django==4.1.3
django-cors-headers==3.13.0
django-storages==1.13.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
# duplicity==0.8.12.0
# entrypoints==0.3
# exceptiongroup==1.0.4
# factory-boy==3.2.1
# Faker==15.3.4
# fasteners==0.14.1
# future==0.18.2
# h11==0.14.0
# httplib2==0.14.0
# idna==2.8
# iniconfig==1.1.1
# jmespath==1.0.1
# keyring==18.0.1
# kombu==5.2.4
# # language-selector==0.1
# launchpadlib==1.10.13
# lazr.restfulclient==0.14.2
# lazr.uri==1.0.3
# lockfile==0.12.2
# # louis==3.12.0
# macaroonbakery==1.3.1
# Mako==1.1.0
# MarkupSafe==1.1.0
# monotonic==1.5
# netifaces==0.10.4
# oauthlib==3.1.0
# olefile==0.46
# outcome==1.2.0
# packaging==21.3
# paramiko==2.6.0
# pexpect==4.6.0
Pillow==7.0.0
# pluggy==1.0.0
# prompt-toolkit==3.0.32
# protobuf==3.6.1
psycopg2-binary==2.9.5
# pycairo==1.16.2
# pycups==1.9.73
# # PyGObject==3.36.0
PyJWT==1.7.1
# pymacaroons==0.13.0
pymemcache==4.0.0
# PyNaCl==1.3.0
# pyparsing==3.0.9
# pyRFC3339==1.1
# PySocks==1.7.1
# pytest==7.2.0
# # python-apt==2.0.0+ubuntu0.20.4.8
# python-dateutil==2.7.3
# # python-debian===0.1.36ubuntu1
# pytz==2022.6
# pyxdg==0.26
# PyYAML==5.3.1
redis==4.3.4
# reportlab==3.5.34
# requests==2.28.1
# requests-unixsocket==0.2.0
# s3transfer==0.6.0
# SecretStorage==2.3.1
selenium==4.7.2
# simplejson==3.16.0
simplejwt==2.0.1
# six==1.14.0
# sniffio==1.3.0
# sortedcontainers==2.4.0
sqlparse==0.4.3
# # systemd-python==234
# tomli==2.0.1
# trio==0.22.0
# trio-websocket==0.9.2
# typing==3.7.4.3
# # ubuntu-advantage-tools==27.12
# # ubuntu-drivers-common==0.0.0
# # ufw==0.36
# # unattended-upgrades==0.1
# urllib3==1.26.13
# # usb-creator==0.3.7
# vine==5.0.0
# wadllib==1.3.3
# wcwidth==0.2.5
# wrapt==1.14.1
# wsproto==1.2.0
# # xkit==0.0.0
